@@ -29,6 +29,7 @@
     "@electron-toolkit/preload": "^3.0.1",
     "@electron-toolkit/utils": "^3.0.0",
     "@tomhuel/jsonhandler": "^1.0.1",
+    "bits-ui": "^1.3.19",
     "decompress": "^4.2.1",
     "electron-updater": "^6.3.9",
     "fs-extra": "^11.2.0",
@@ -38,7 +39,7 @@
     "@electron-toolkit/eslint-config-prettier": "^2.0.0",
     "@electron-toolkit/eslint-config-ts": "^2.0.0",
     "@electron-toolkit/tsconfig": "^1.0.1",
-    "@sveltejs/vite-plugin-svelte": "^3.1.1",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0-next.6",
     "@types/node": "^20.14.8",
     "autoprefixer": "^10.4.19",
     "electron": "^31.0.2",
@@ -49,8 +50,8 @@
     "postcss": "^8.4.39",
     "prettier": "^3.3.2",
     "prettier-plugin-svelte": "^3.2.5",
-    "svelte": "^4.2.18",
-    "svelte-check": "^3.8.2",
+    "svelte": "^5.28.2",
+    "svelte-check": "^4.1.6",
     "tailwind": "^4.0.0",
     "tailwindcss": "^3.4.4",
     "tslib": "^2.6.3",
