@@ -1,54 +1,62 @@
 <script lang="ts">
-  import { createEventDispatcher } from "svelte";
-  import Checkbox from "./checkbox.svelte";
-  import ChangeNameModal from "../modal/changeNameModal/changeNameModal.svelte";
+	import Checkbox from "./checkbox.svelte";
+	import ChangeNameModal from "../modal/changeNameModal/changeNameModal.svelte";
 
-  const openDir = async (path) => {
-    await window.electron.openDir(path);
-  };
+	let {
+		modName = "",
+		realModName = "",
+		active = false,
+		path = "",
+		reload,
+	}: {
+		modName: string;
+		realModName: string;
+		active: boolean;
+		path: string;
+		reload: () => void;
+	} = $props();
 
-  const deleteMod = async (realModName) => {
-    await window.electron.deleteMod(realModName);
-    dispatch("reload");
-  };
+	const openDir = async (path) => {
+		await window.electron.openDir(path);
+	};
 
-  const enableMod = async () => {
-    await window.electron.enableMod(realModName);
-  };
+	const deleteMod = async (realModName) => {
+		await window.electron.deleteMod(realModName);
+		reload();
+	};
 
-  const disableMod = async () => {
-    await window.electron.disableMod(realModName);
-  };
+	const enableMod = async () => {
+		console.log("enabling mod", realModName);
+		await window.electron.enableMod(realModName);
+	};
 
-  export let modName: string = "";
-  export let realModName: string = "";
-  export let active: boolean = false;
-  export let path: string = "";
-
-  const dispatch = createEventDispatcher();
+	const disableMod = async () => {
+		console.log("disabling mod", realModName);
+		await window.electron.disableMod(realModName);
+	};
 </script>
 
 <div class="border border-light bg-gray rounded-button py-3 px-4 flex flex-col">
-  <div class="flex justify-between gap-2">
-    <h3 class=" text-light text-3xl font-yeon truncate">{modName}</h3>
-    <Checkbox
-      {active}
-      id={realModName}
-      on:enable={enableMod}
-      on:disable={disableMod}
-    />
-  </div>
-  <button
-    on:click={async () => await openDir(path)}
-    class="text-start text-drown-gray text-xl font mb-3 truncate"
-    >{realModName}</button
-  >
-  <!-- Open File System Window -->
-  <div class="flex justify-between">
-    <ChangeNameModal name={modName} realname={realModName} on:reload />
-    <button
-      class=" text-primary font-abz hover:underline"
-      on:click={async () => await deleteMod(realModName)}>Delete</button
-    >
-  </div>
+	<div class="flex justify-between gap-2">
+		<h3 class=" text-light text-3xl font-yeon truncate">{modName}</h3>
+		<Checkbox
+			{active}
+			id={realModName}
+			enable={enableMod}
+			disable={disableMod}
+		/>
+	</div>
+	<button
+		onclick={async () => await openDir(path)}
+		class="text-start text-drown-gray text-xl font mb-3 truncate"
+		>{realModName}</button
+	>
+	<!-- Open File System Window -->
+	<div class="flex justify-between">
+		<ChangeNameModal name={modName} realname={realModName} reload />
+		<button
+			class=" text-primary font-abz hover:underline"
+			onclick={async () => await deleteMod(realModName)}>Delete</button
+		>
+	</div>
 </div>
