@@ -1,8 +1,12 @@
 <script lang="ts">
-	export let buttonClassname: string = "";
-	export let title: string;
-	export let modalBodyClassname: string = "p-5";
-	export let displayModal = false;
+	let {
+		children,
+		buttonContent,
+		buttonClassname = "",
+		title,
+		modalBodyClassname = "p-5",
+		displayModal = false,
+	} = $props();
 
 	const closeModal = () => {
 		displayModal = false;
@@ -16,17 +20,33 @@
 </script>
 
 <!-- Button -->
-<button on:click={openModal} class={buttonClassname} {title}>
-    <slot name="buttonContent" />
+<button onclick={openModal} class={buttonClassname} {title}>
+	{#if buttonContent}
+		{@render buttonContent()}
+	{:else}
+		<span>No name found!</span>
+	{/if}
 </button>
 <!-- Modal Body -->
-<div class="flex justify-center items-center fixed min-h-screen w-screen z-10 opacity-65 top-0 left-0 bg-darker p-5 {displayModal ? '' : 'hidden'}" />
-<div class="fixed top-0 left-0 flex h-screen w-screen justify-center items-center opacity-100 z-20 {displayModal ? '' : 'hidden'}">
+<div
+	class="flex justify-center items-center fixed min-h-screen w-screen z-10 opacity-65 top-0 left-0 bg-darker p-5 {displayModal
+		? ''
+		: 'hidden'}"
+></div>
+<div
+	class="fixed top-0 left-0 flex h-screen w-screen justify-center items-center opacity-100 z-20 {displayModal
+		? ''
+		: 'hidden'}"
+>
 	<div class=" bg-dark rounded-button border border-gray flex flex-col z-20">
-		<div class="h-full flex flex-col justify-center items-center min-w-96 max-h-96">
-			<div class="flex p-5 border-b border-light w-full justify-between items-center">
+		<div
+			class="h-full flex flex-col justify-center items-center min-w-96 max-h-96"
+		>
+			<div
+				class="flex p-5 border-b border-light w-full justify-between items-center"
+			>
 				<span class="text-light text-2xl font-abz">{title}</span>
-				<button on:click={closeModal}>
+				<button onclick={closeModal} aria-label="Close modal">
 					<svg
 						xmlns="http://www.w3.org/2000/svg"
 						width="36"
@@ -42,7 +62,7 @@
 				</button>
 			</div>
 			<div class="{modalBodyClassname} w-full overflow-y-scroll">
-				<slot />
+				{@render children()}
 			</div>
 		</div>
 	</div>
