@@ -1,4 +1,4 @@
-<script>
+<script lang="ts">
 	import Navbar from "./components/navbar/Navbar.svelte";
 	import Modcard from "./components/card/modcard.svelte";
 	import { onMount } from "svelte";
@@ -8,16 +8,16 @@
 		mods = await window.electron.getMods();
 	};
 
-	let mods = [];
+	let mods = $state([]);
 
-	let newUpdate = false;
+	let newUpdate = $state(false);
 
-	let search = "";
+	let search = $state("");
 
 	onMount(async () => {
 		document.addEventListener("updateMessage", () => {
 			newUpdate = true;
-			console.log('new update!');
+			console.log("new update!");
 		});
 
 		await getMods();
@@ -39,7 +39,7 @@
 		</div>
 	</main>
 {:else}
-	<Navbar on:search={filterMods} />
+	<Navbar searchEvent={filterMods} />
 	<main class="px-6 py-5">
 		<div
 			class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 2xl:gap-x-20 gap-8"
@@ -50,7 +50,7 @@
 					active={mod.status === "ENABLED"}
 					realModName={mod.realname}
 					path={mod.path}
-					on:reload={filterMods}
+					reload={filterMods}
 				/>
 			{/each}
 		</div>
