@@ -2,7 +2,7 @@
     import Reset from "../../../icons/reset.svelte";
 </script>
 
-<button class="bg-primary flex justify-between items-center gap-2 text-light ps-2 pe-3 rounded-button my-3 transition-all duration-500 hover:bg-primary-darker" on:click>
+<button class="bg-primary flex justify-between items-center gap-2 text-light ps-2 pe-3 rounded-button my-3 transition-all duration-500 hover:bg-primary-darker">
     <Reset className="w-6" />
     <span class="font-yeon text-lg leading-10 py-1">Fix Models</span>
 </button>
